refactor(Btn): Change Btn Name from AddToPlaylist to AddToFavourites
Removed AddToPlaylist component and replaced it with AddToFavourites in multiple files
